Mudando metodo de fechar ticket para modificar data de fechamento para momento atual do fechamento, adicionando metodo para obter tickets por mes fechados e contar tickets por status para dashboard

+++ src/main/java/br/ufma/glp/unidesk/backend/api/v1/dto/model/DashboardModel.java
@@ -16,8 +16,8 @@ public class DashboardModel {
     Long totalTicketsAbertos;
     Long totalTicketsPendentes;
     Long totalTicketsEmAndamento;
-    Long porcentagemProgresso;
+    Double porcentagemProgresso;
     Double porcentagemAbertos;
-    Long porcentagemResolvidos;
-    Long porcentagemAndamento;
+    Double porcentagemResolvidos;
+    Double porcentagemAndamento;
 }
